Add default.sh
If present, default.sh should be run when you run just '.' as a command.

The . file.ext means "do the default action to file.ext given its type".
It's kind of like double-clicking on a file via Desktop interface.  I
interpret lone . to mean "do the default action for this directory".
Now, the default direction of '. <dir>' is to cd to <dir>.  But moving
to the same dir is pointless, so we can do something more elaborate like
run default.sh (if present).

Also, I'm thinking of having a default default action for directories
that don't have a default.sh.