Use line buffer to parse content length to prevent hang on last recv()
Yeah, converting the raw buffer reads into a sequence of lines and then
parsing the content length out of the HTTP header is necessary otherwise
you hang indefinitely on the last blocking recv() call.

Of course, the example from the book did not even try to deal with these
unfortunate realities.